+++ src/cmd/trace/goroutines.go
@@ -169,7 +169,7 @@ func httpGoroutine(w http.ResponseWriter, r *http.Request) {
 var templGoroutine = template.Must(template.New("").Funcs(template.FuncMap{
 	"prettyDuration": func(nsec int64) template.HTML {
 		d := time.Duration(nsec) * time.Nanosecond
-		return template.HTML(niceDuration(d))
+		return template.HTML(d.String())
 	},
 	"percent": func(dividend, divisor int64) template.HTML {
 		if divisor == 0 {
